third_party/blink/web_tests/external/wpt/css/css-page/page-rule-declarations-003-expected.txt

This is a testharness.js-based test.
[FAIL] CSS Paged Media: parsing @page selectors
  assert_equals: missing @page selectors expected 0 but got 7
Harness: the test ran to completion.
